+++ lib/rooibos/message.rb
@@ -91,6 +91,7 @@ module Rooibos
 end
 
 require_relative "message/timer"
+require_relative "message/open"
 require_relative "message/http_response"
 require_relative "message/system/batch"
 require_relative "message/system/stream"